⚙️ Move Engine's to_prepare to an initializer
This commit will resolve load ordering.  Previously, we were loading the
decorators from HykuKnapsack before we load the App's, which is standard
behavior, but we actually want the decorators from HykuKnapsack to load
last.
