+++ app/presenters/adl/collection_presenter_decorator.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
 # OVERRIDE HYRAX 3.5.0
-module Hyrax
+module Adl
   module CollectionPresenterDecorator
     module ClassMethods
       # Terms is the list of fields displayed by
       # app/views/collections/_show_descriptions.html.erb
       def terms
-        super + Collection.additional_terms
+        super - [:size] + Collection.additional_terms
       end
     end
   end
 end
 
-Hyrax::CollectionPresenter.singleton_class.prepend(Hyrax::CollectionPresenterDecorator::ClassMethods)
+Hyrax::CollectionPresenter.singleton_class.prepend(Adl::CollectionPresenterDecorator::ClassMethods)
