+++ spec/presenters/adl/collection_presenter_decorator_spec.rb
@@ -23,8 +23,7 @@ RSpec.describe Hyrax::CollectionPresenter do
   describe '#terms_with_values' do
     let(:instance) { described_class.new(document, nil) }
 
-    # TODO: this spec fails because Hyku has an override by the same name
-    xit 'includes values from collection' do
+    it 'includes values from collection' do
       expect(instance.terms_with_values).to include(:official_url) # Which is a collection property.
     end
   end
